Added package to verify network connectivity and rename package core/platform to core/network
