+++ Flutter/vs_code_app/lib/core/network/network_info.dart
