+++ Flutter/vs_code_app/lib/features/number_trivia/data/repositories/number_trivia_repository_impl.dart
@@ -1,7 +1,7 @@
 import 'package:dartz/dartz.dart';
 import 'package:vs_code_app/core/error/exception.dart';
 import 'package:vs_code_app/core/error/failures.dart';
-import 'package:vs_code_app/core/platform/network_info.dart';
+import 'package:vs_code_app/core/network/network_info.dart';
 import 'package:vs_code_app/features/number_trivia/data/datasources/number_trivia_local_data_source.dart';
 import 'package:vs_code_app/features/number_trivia/data/datasources/number_trivia_remote_data_source.dart';
 import 'package:vs_code_app/features/number_trivia/data/models/number_trivia_model.dart';
