+++ Flutter/vs_code_app/pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   http: ^0.12.0+2
   # Local cache
   shared_preferences: ^0.5.3+4
+  data_connection_checker: ^0.3.4
 
 dev_dependencies:
   flutter_test:
