+++ Flutter/vs_code_app/test/features/number_trivia/data/repositories/number_trivia_repository_impl_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:vs_code_app/core/error/exception.dart';
 import 'package:vs_code_app/core/error/failures.dart';
-import 'package:vs_code_app/core/platform/network_info.dart';
+import 'package:vs_code_app/core/network/network_info.dart';
 import 'package:vs_code_app/features/number_trivia/data/datasources/number_trivia_local_data_source.dart';
 import 'package:vs_code_app/features/number_trivia/data/datasources/number_trivia_remote_data_source.dart';
 import 'package:vs_code_app/features/number_trivia/data/models/number_trivia_model.dart';
